pr feedback, migrated detection_to_object to module, named imports
